requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.0.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
elevenlabs==1.56.0
frozenlist==1.5.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==2.10
incremental==24.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.2.0
ngrok-api==0.13.0
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==7.2.3
pyOpenSSL==25.0.0
python-dotenv==1.0.0
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==14.0.0
service-identity==24.2.0
setuptools==78.1.0
sniffio==1.3.1
sqlparse==0.5.3
twilio==8.10.3
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==1.26.20
uvicorn==0.34.0
websockets==15.0.1
yarl==1.18.3
zope.interface==7.2
